2005-11-15  Lillian Angel  <langel@redhat.com>
        * javax/swing/plaf/basic/BasicLookAndFeel.java
        (initComponentDefaults): Default rowHeight for tree's should
        be 0. This is Sun and IBM's default.
        * javax/swing/plaf/metal/MetalLookAndFeel.java
        (initComponentDefaults): Default rowHeight for tree's should
        e 0. This is Sun and IBM's default.
        * javax/swing/plaf/basic/BasicTreeUI.java:
        Added a default rowHeight field.
        (setRowHeight): Set the rowHeight to the class default if parameter
        is 0. Tree row height should never be set to 0. 20 is the minimum.
        (installDefaults): fixed call to tree.setRowHeight
        (shouldPaintExpandControl): Added to check to prevent NPE.

+++ javax/swing/plaf/metal/MetalLookAndFeel.java
@@ -1162,7 +1162,7 @@ public class MetalLookAndFeel extends BasicLookAndFeel
       "Tree.line", getPrimaryControl(),
       "Tree.openIcon", MetalIconFactory.getTreeFolderIcon(),
       "Tree.rightChildIndent", new Integer(13),
-      "Tree.rowHeight", new Integer(20),
+      "Tree.rowHeight", new Integer(0),
       "Tree.scrollsOnExpand", Boolean.TRUE,
       "Tree.selectionBackground", getTextHighlightColor(),
       "Tree.selectionBorder", new BorderUIResource.LineBorderUIResource(new Color(102, 102, 153)),
